haptics work, the seconds hand gets set to the right value when the view loads
this is important to the user experience because it allows the user to tell the seconds. this allows the user to be more precise with their time telling, and being precise is always (usually) a good thing. this adds positively to the user experience, because even though it's another thing (more clutter), it's relatively unintrusive, meaning we've stuck to the minimalistic design, while adding more information. this is, if you ask me, quite a feat, considering that adding more information generally has the tendency to increase clutter and *detract* from the user experience. if only markdown worked in commit messages.